+++ src/library/DIPS.Mobile.UI/Components/Buttons/iOS/ButtonHandler.cs
@@ -1,4 +1,5 @@
 using CoreGraphics;
+using DIPS.Mobile.UI.API.Library;
 using DIPS.Mobile.UI.Platforms.iOS;
 using Microsoft.Maui.Handlers;
 using Microsoft.Maui.Platform;
@@ -9,20 +10,12 @@ namespace DIPS.Mobile.UI.Components.Buttons;
 
 public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
 {
-    private int LargeButtonHeight { get; } = Sizes.GetSize(SizeName.size_12);
+    
     
     protected override UIButton CreatePlatformView()
     {
         var button = new UIButtonWithExtraTappableArea();
         return button;
-    }
-
-    protected override void ConnectHandler(UIButton platformView)
-    {
-        base.ConnectHandler(platformView);
-
-        var percentOfHeight = (float)VirtualView.Height / LargeButtonHeight;
-        platformView.ImageView.Transform = CGAffineTransform.MakeScale(percentOfHeight, percentOfHeight);
     }
 
     private partial void AppendPropertyMapper()
@@ -45,7 +38,7 @@ public partial class ButtonHandler : Microsoft.Maui.Handlers.ButtonHandler
     {
         await MapImageSourceAsync(handler, button);
         MapImageTintColor(handler, button);
-    }
+        }
 
     private static partial void MapImageTintColor(ButtonHandler handler, Button button)
     {