+++ src/library/DIPS.Mobile.UI/Components/ContextMenus/ContextMenuItem.cs
@@ -3,7 +3,7 @@ namespace DIPS.Mobile.UI.Components.ContextMenus;
 /// <summary>
 /// A context menu item to use in a context menu
 /// </summary>
-public partial class ContextMenuItem : Element
+public partial class ContextMenuItem : Element, IContextMenuItem
 {
     internal void SendClicked(ContextMenu contextMenu)
     {