+++ src/library/DIPS.Mobile.UI/Components/Labels/Label.cs
@@ -1,3 +1,6 @@
+using DIPS.Mobile.UI.Resources.Styles;
+using DIPS.Mobile.UI.Resources.Styles.Label;
+
 namespace DIPS.Mobile.UI.Components.Labels
 {
     public partial class Label : Microsoft.Maui.Controls.Label
@@ -6,6 +9,7 @@ namespace DIPS.Mobile.UI.Components.Labels
         {
             this.SetAppThemeColor(TextColorProperty, ColorName.color_neutral_90);
             MaxLines = int.MaxValue;
+            Style = Styles.GetLabelStyle(LabelStyle.Body300);
         }
     }
 }