+++ src/library/DIPS.Mobile.UI/Components/TextFields/Entry.cs
@@ -1,0 +1,6 @@
+namespace DIPS.Mobile.UI.Components.TextFields;
+
+public partial class Entry : Microsoft.Maui.Controls.Entry
+{
+    
+}