+++ src/library/DIPS.Mobile.UI/Effects/Touch/iOS/TouchPlatformEffect.cs
@@ -56,8 +56,10 @@ public partial class TouchPlatformEffect
     {
         if (m_touchMode is Touch.TouchMode.Tap or Touch.TouchMode.Both)
         {
-            if(Control.GestureRecognizers != null)
+            if (Control.GestureRecognizers != null)
+            {
                 Control.RemoveGestureRecognizer(m_tapGestureRecognizer!);
+            }
         }
 
         if (m_touchMode is Touch.TouchMode.LongPress or Touch.TouchMode.Both)